Bugfix: Add default scale to text objects Change: Uses floats where Tiku make the api use floats. Change: Added more parsing to text objects to handle br and tags that close themselves.

+++ AccSaber/Models/AccSaberLeaderboardEntry.cs
@@ -38,7 +38,7 @@ namespace AccSaber.Models
         public int BadCuts { get; set; }
 
         [JsonProperty("baseXp")]
-        public int BaseXp { get; set; }
+        public float BaseXp { get; set; }
 
         [JsonProperty("blScoreId")]
         public string? BlScoreId { get; set; }
